Add minimal FRAM driver.

+++ src/board/spi.h
@@ -77,7 +77,7 @@ uint8_t SPI_write8(uint8_t spi_num, uint8_t addr, uint8_t *data);
  * @param len Number of bytes to be read
  * @returns Status byte
  */
-uint8_t SPI_burstread8(uint8_t spi_num, uint8_t addr, uint8_t *data, uint16_t len);
+uint8_t SPI_burstread8(uint8_t spi_num, uint8_t addr, uint8_t *data, uint32_t len);
 
 /**
  * @function
@@ -91,7 +91,7 @@ uint8_t SPI_burstread8(uint8_t spi_num, uint8_t addr, uint8_t *data, uint16_t le
  * @param len Number of bytes to be written
  * @returns Status byte
  */
-uint8_t SPI_burstwrite8(uint8_t spi_num, uint8_t addr, uint8_t *data, uint16_t len);
+uint8_t SPI_burstwrite8(uint8_t spi_num, uint8_t addr, uint8_t *data, uint32_t len);
 
 /**
  * @function
@@ -131,7 +131,7 @@ uint8_t SPI_write16(uint8_t spi_num, uint16_t addr, uint8_t *data);
  * @param len Number of bytes to be read
  * @returns Status byte
  */
-uint8_t SPI_burstread16(uint8_t spi_num, uint16_t addr, uint8_t *data, uint16_t len);
+uint8_t SPI_burstread16(uint8_t spi_num, uint16_t addr, uint8_t *data, uint32_t len);
 
 /**
  * @function
@@ -145,7 +145,7 @@ uint8_t SPI_burstread16(uint8_t spi_num, uint16_t addr, uint8_t *data, uint16_t 
  * @param len Number of bytes to be written
  * @returns Status byte
  */
-uint8_t SPI_burstwrite16(uint8_t spi_num, uint16_t addr, uint8_t *data, uint16_t len);
+uint8_t SPI_burstwrite16(uint8_t spi_num, uint16_t addr, uint8_t *data, uint32_t len);
 
 /**
  * @function
@@ -185,7 +185,7 @@ uint8_t SPI_write32(uint8_t spi_num, uint32_t addr, uint8_t *data);
  * @param len Number of bytes to be read
  * @returns Status byte
  */
-uint8_t SPI_burstread32(uint8_t spi_num, uint32_t addr, uint8_t *data, uint16_t len);
+uint8_t SPI_burstread32(uint8_t spi_num, uint32_t addr, uint8_t *data, uint32_t len);
 
 /**
  * @function
@@ -199,6 +199,6 @@ uint8_t SPI_burstread32(uint8_t spi_num, uint32_t addr, uint8_t *data, uint16_t 
  * @param len Number of bytes to be written
  * @returns Status byte
  */
-uint8_t SPI_burstwrite32(uint8_t spi_num, uint32_t addr, uint8_t *data, uint16_t len);
+uint8_t SPI_burstwrite32(uint8_t spi_num, uint32_t addr, uint8_t *data, uint32_t len);
 
 #endif /*  __SPI_H__ */
